hook생성
